chore(server): Move server io stream to separate module
